pyproject.toml
[project]
name = "filldoc"
version = "0.1.0"
description = "Fill DOCX templates with data from configuration files"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-docx>=1.2.0",
    "pyyaml>=6.0.2",
    "tomli>=2.2.1",
]

[project.scripts]
filldoc = "filldoc:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
